Update icons.
Previously, I had converted all the png icon files to pdf because there
was a problem with the background of the pngs.  As I recall, the background
color was totally transparent but black, which caused the icons to be
printed with black boarders on some builds.  Either because of changes in
LaTeX or my build, that no longer seems to be a problem, so I am just
copying the .png files.